requirements.txt
# SpendSense API Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pandas>=2.2.0
faker>=19.12.0
pytest>=7.4.3
pytest-asyncio>=0.21.0
pydantic>=2.5.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
openai>=1.0.0
guardrails-ai>=0.6.0
PyJWT>=2.8.0